+++ elkm1_lib/panel.py
@@ -1,16 +1,18 @@
 """Definition of an ElkM1 Area"""
 import datetime as dt
 import time
+from typing import Any, Optional
 
 from .const import ElkRPStatus
 from .elements import Element
+from .elk import Elk
 from .message import lw_encode, rw_encode, sp_encode, ss_encode, sw_encode, vn_encode
 
 
 class Panel(Element):
     """Class representing the overall Elk panel"""
 
-    def __init__(self, elk):
+    def __init__(self, elk: Elk) -> None:
         super().__init__(0, elk)
         self.real_time_clock = None
         self.elkm1_version = None
@@ -30,37 +32,37 @@ class Panel(Element):
         self._elk.add_handler("SS", self._ss_handler)
         self._elk.add_handler("UA", self._ua_handler)
 
-    def sync(self):
+    def sync(self) -> None:
         """Retrieve panel information from ElkM1"""
         self._elk.send(vn_encode())
         self._elk.send(lw_encode())
         self._elk.send(ss_encode())
         # Don't sync UA from here as it is used as a "sync complete" marker
 
-    def speak_word(self, word):
+    def speak_word(self, word: int) -> None:
         """(Helper) Speak word."""
         self._elk.send(sw_encode(word))
 
-    def speak_phrase(self, phrase):
+    def speak_phrase(self, phrase: int) -> None:
         """(Helper) Speak phrase."""
         self._elk.send(sp_encode(phrase))
 
-    def set_time(self, datetime=None):
+    def set_time(self, datetime: Optional[dt.datetime] = None) -> None:
         """(Helper) Set the time given a datetime."""
         if datetime is None:
             struct_time = time.localtime()
             datetime = dt.datetime(*struct_time[:6])
         self._elk.send(rw_encode(datetime))
 
-    def _vn_handler(self, elkm1_version, xep_version):
+    def _vn_handler(self, elkm1_version: str, xep_version: str) -> None:
         self.setattr("elkm1_version", elkm1_version, False)
         self.setattr("xep_version", xep_version, True)
 
-    def _xk_handler(self, real_time_clock):
+    def _xk_handler(self, real_time_clock: str) -> None:
         self.setattr("real_time_clock", real_time_clock, True)
 
-    def _ss_handler(self, system_trouble_status):
-        def _get_status(index, trouble, zone_encoded=False):
+    def _ss_handler(self, system_trouble_status: str) -> None:
+        def _get_status(index: int, trouble: str, zone_encoded: bool = False) -> None:
             if system_trouble_status[index] != "0":
                 if zone_encoded:
                     zone = ord(system_trouble_status[index]) - 0x30
@@ -68,7 +70,7 @@ class Panel(Element):
                 else:
                     statuses.append(trouble)
 
-        statuses = []
+        statuses: list[str] = []
         _get_status(0, "AC Fail")
         _get_status(1, "Box Tamper", True)
         _get_status(2, "Fail To Communicate")
@@ -94,13 +96,13 @@ class Panel(Element):
         _get_status(33, "Fire", True)
         self.setattr("system_trouble_status", ", ".join(statuses), True)
 
-    def _rp_handler(self, remote_programming_status):
+    def _rp_handler(self, remote_programming_status: int) -> None:
         if remote_programming_status == ElkRPStatus.DISCONNECTED.value:
             self._elk.resume()
         else:
             self._elk.pause()
         self.setattr("remote_programming_status", remote_programming_status, True)
 
-    def _ua_handler(self, **kwargs):
+    def _ua_handler(self, **kwargs: dict[str, Any]) -> None:
         self.setattr("user_code_length", kwargs["user_code_length"], False)
         self.setattr("temperature_units", kwargs["temperature_units"], True)