+++ elkm1_lib/users.py
@@ -2,21 +2,31 @@
 
 from .const import Max, TextDescriptions
 from .elements import Element, Elements
+from .elk import Elk
 
 
 class User(Element):
     """Class representing an User"""
 
-    def __init__(self, index, elk):  # pylint: disable=useless-super-delegation
-        super().__init__(index, elk)
-
 
 class Users(Elements):
     """Handling for multiple areas"""
 
-    def __init__(self, elk):
+    def __init__(self, elk: Elk) -> None:
         super().__init__(elk, User, Max.USERS.value)
 
-    def sync(self):
+    def sync(self) -> None:
         """Retrieve areas from ElkM1"""
         self.get_descriptions(TextDescriptions.USER.value)
+
+    def username(self, user_number: int) -> str:
+        """Return name of user."""
+        if 0 <= user_number < self.max_elements:
+            return self.elements[user_number].name
+        if user_number == 201:
+            return "*Program*"
+        if user_number == 202:
+            return "*Elk RP*"
+        if user_number == 203:
+            return "*Quick arm*"
+        return ""