+++ elkm1_lib/util.py
@@ -1,5 +1,7 @@
 """Utility functions"""
 
+from __future__ import annotations
+
 import ssl
 
 TLS_VERSIONS = {
@@ -16,50 +18,39 @@ TLS_VERSIONS = {
 }
 
 
-def url_scheme_is_secure(url):
+def url_scheme_is_secure(url: str) -> bool:
     """Check if the URL is one that requires SSL/TLS."""
     scheme, _dest = url.split("://")
     return scheme.startswith("elks")
 
 
-def parse_url(url):
-    """Parse a Elk connection string """
+def parse_url(url: str) -> tuple[str, str, int, ssl.SSLContext | None]:
+    """Parse a Elk connection string"""
     scheme, dest = url.split("://")
     host = None
     ssl_context = None
     if scheme == "elk":
-        host, port = dest.split(":") if ":" in dest else (dest, 2101)
+        host, port = dest.split(":") if ":" in dest else (dest, "2101")
     elif TLS_VERSIONS.get(scheme):
-        host, port = dest.split(":") if ":" in dest else (dest, 2601)
+        host, port = dest.split(":") if ":" in dest else (dest, "2601")
         ssl_context = ssl.SSLContext(ssl.PROTOCOL_TLS)
-        ssl_context.minimum_version = TLS_VERSIONS.get(scheme)
-        ssl_context.maximum_version = TLS_VERSIONS.get(scheme)
+        if tls := TLS_VERSIONS.get(scheme):
+            ssl_context.minimum_version = tls
+            ssl_context.maximum_version = tls
+
         ssl_context.verify_mode = ssl.CERT_NONE
         scheme = "elks"
     elif scheme == "serial":
-        host, port = dest.split(":") if ":" in dest else (dest, 115200)
+        host, port = dest.split(":") if ":" in dest else (dest, "115200")
     else:
-        raise ValueError("Invalid scheme '%s'" % scheme)
+        raise ValueError(f"Invalid scheme '{scheme}'")
     return (scheme, host, int(port), ssl_context)
 
 
-def pretty_const(value):
+def pretty_const(value: str) -> str:
     """Make a constant pretty for printing in GUI"""
     words = value.split("_")
     pretty = words[0].capitalize()
     for word in words[1:]:
         pretty += " " + word.lower()
     return pretty
-
-
-def username(elk, user_number):
-    """Return name of user."""
-    if 0 <= user_number < elk.users.max_elements:
-        return elk.users[user_number].name
-    if user_number == 201:
-        return "*Program*"
-    if user_number == 202:
-        return "*Elk RP*"
-    if user_number == 203:
-        return "*Quick arm*"
-    return ""